# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: School Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /exams:
    get:
      operationId: listExams
      tags: [exams]
      parameters:
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: subjectId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
    post:
      operationId: createExam
      tags: [exams]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /exams/{id}:
    get:
      operationId: getExam
      tags: [exams]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    patch:
      operationId: updateExam
      tags: [exams]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExamUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    delete:
      operationId: deleteExam
      tags: [exams]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Results ───────────────────────────────────────────────────────────
  /results:
    get:
      operationId: listResults
      tags: [results]
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: examId
          in: query
          schema:
            type: ["integer", "null"]
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Result"
    post:
      operationId: createResult
      tags: [results]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResultInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"
  /results/{id}:
    patch:
      operationId: updateResult
      tags: [results]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResultUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Result"

  # ── Fees ──────────────────────────────────────────────────────────────
  /fees:
    get:
      operationId: listFees
      tags: [fees]
      parameters:
        - name: studentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
            enum: [paid, unpaid, partial, overdue]
      responses:
        "200":
          description: Fee records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fee"
    post:
      operationId: createFee
      tags: [fees]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fee"
  /fees/{id}:
    patch:
      operationId: updateFee
      tags: [fees]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fee"
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      parameters:
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeave
      tags: [leaves]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}:
    patch:
      operationId: updateLeave
      tags: [leaves]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"

  # ── Dashboard / Analytics (wow endpoints) ─────────────────────────────
...
components:
  schemas:
...
    Exam:
      type: object
      required: [id, title, classId, subjectId, examDate, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        classId:
          type: integer
        className:
          type: ["string", "null"]
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        examDate:
          type: string
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        totalMarks:
          type: integer
        passingMarks:
          type: integer
        status:
          type: string
          enum: [scheduled, ongoing, completed, cancelled]
        venue:
          type: ["string", "null"]
        createdAt:
          type: string
    ExamInput:
      type: object
      required: [title, classId, subjectId, examDate, totalMarks, passingMarks]
      properties:
        title:
          type: string
        classId:
          type: integer
        subjectId:
          type: integer
        examDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        totalMarks:
          type: integer
        passingMarks:
          type: integer
        venue:
          type: string
    ExamUpdate:
      type: object
      properties:
        title:
          type: string
        examDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        totalMarks:
          type: integer
        passingMarks:
          type: integer
        status:
          type: string
          enum: [scheduled, ongoing, completed, cancelled]
        venue:
          type: string

    # Result
    Result:
      type: object
      required: [id, examId, studentId, studentName, marksObtained, totalMarks, createdAt]
      properties:
        id:
          type: integer
        examId:
          type: integer
        examTitle:
          type: ["string", "null"]
        studentId:
          type: integer
        studentName:
          type: string
        marksObtained:
          type: number
        totalMarks:
          type: integer
        grade:
          type: ["string", "null"]
        remarks:
          type: ["string", "null"]
        isPassed:
          type: boolean
        createdAt:
          type: string
    ResultInput:
      type: object
      required: [examId, studentId, marksObtained]
      properties:
        examId:
          type: integer
        studentId:
          type: integer
        marksObtained:
          type: number
        remarks:
          type: string
    ResultUpdate:
      type: object
      properties:
        marksObtained:
          type: number
        remarks:
          type: string

    # Fee
    Fee:
      type: object
      required: [id, studentId, studentName, amount, dueDate, status, feeType, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        amount:
          type: number
        paidAmount:
          type: number
        dueDate:
          type: string
        paidDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [paid, unpaid, partial, overdue]
        feeType:
          type: string
          enum: [tuition, transport, library, sports, examination, other]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    FeeInput:
      type: object
      required: [studentId, amount, dueDate, feeType]
      properties:
        studentId:
          type: integer
        amount:
          type: number
        dueDate:
          type: string
        feeType:
          type: string
          enum: [tuition, transport, library, sports, examination, other]
        description:
          type: string
    FeeUpdate:
      type: object
      properties:
        paidAmount:
          type: number
        paidDate:
          type: string
        status:
          type: string
          enum: [paid, unpaid, partial, overdue]
...
    LeaveRequest:
      type: object
      required: [id, userId, userName, userRole, fromDate, toDate, reason, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userRole:
          type: string
          enum: [student, teacher]
        fromDate:
          type: string
        toDate:
          type: string
        reason:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        adminRemarks:
          type: ["string", "null"]
        createdAt:
          type: string
    LeaveInput:
      type: object
      required: [userId, fromDate, toDate, reason]
      properties:
        userId:
          type: integer
        fromDate:
          type: string
        toDate:
          type: string
        reason:
          type: string
    LeaveUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        adminRemarks:
          type: string

    # Dashboard/Analytics
...