# lib/api-zod/src/generated/types/examUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * School Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { ExamUpdateStatus } from './examUpdateStatus';

export interface ExamUpdate {
  title?: string;
  examDate?: string;
  startTime?: string;
  endTime?: string;
  totalMarks?: number;
  passingMarks?: number;
  status?: ExamUpdateStatus;
  venue?: string;
}